feat(BUX-291): initial refactorization. Split model_sync_transactions.go into smaller, more specialized files
